+++ packages/shared/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./types";
 export * from "./utils";
 export * from "./file-types";
+export * from "./error";